backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
django-environ==0.12.0
